+++ mosapi/getStateResponse.go
@@ -20,7 +20,7 @@ func (c *Client) GetStateResponse(ctx context.Context) (*StateResponse, error) {
 	if err != nil {
 		return nil, err
 	}
-	resp, err := c.Do(req)
+	resp, err := c.do(req)
 	if err != nil {
 		return nil, err
 	}